rename existing actors
